# application/services/custom_services/geppayments/src/models/entitymodels/geppayments.ts
export interface geppayments 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   user_name: { type: String },
   total_price: { type: String },
   cardholder_name: { type: String },
   invoiceid: { type: String }
}